+++ tecina-api/database/migrations/2018_04_19_182252_create_categories_dishes_table.php
@@ -14,8 +14,8 @@ class CreateCategoriesDishesTable extends Migration
     public function up()
     {
         Schema::create('categories_dishes', function (Blueprint $table) {
-            $table->integer('id_category')->unsigned();
-            $table->integer('id_dish')->unsigned();
+            $table->smallInteger('id_category')->unsigned();
+            $table->smallInteger('id_dish')->unsigned();
 
             $table->primary(['id_category', 'id_dish']);
