+++ tecina-api/database/migrations/2018_06_20_074340_create_wine_age_translations_table.php
@@ -15,8 +15,8 @@ class CreateWineAgeTranslationsTable extends Migration
     {
         Schema::create('wine_age_translations', function (Blueprint $table) {
             $table->smallIncrements('id');
-            $table->integer('wine_age_id')->unsigned();
-            $table->integer('language_id')->unsigned();
+            $table->smallInteger('wine_age_id')->unsigned();
+            $table->smallInteger('language_id')->unsigned();
             $table->foreign('wine_age_id')
                 ->references('id')->on('wine_ages')
                 ->onDelete('cascade');
